requirements.txt
numpy>=1.24
pandas>=2.0
psycopg2-binary>=2.9
yfinance>=0.2
pytest>=9.0
